weapon collision and keyborad (not finished)
