+++ classes/Cloud.py
@@ -26,27 +26,4 @@ class Cloud:
     screen.blit(self.image, (self.x, self.y))
 
 
-class Weapon:
 
-  WEAPON = pygame.image.load(os.path.join(
-      IMAGE_PATH, "5. Weapons/Weapon1_Logo.png"))
-
-  def __init__(self, width):
-    self.x = width + random.randint(500, 1000)
-    self.y = 100
-    self.image = self.WEAPON
-    self.width = self.image.get_width()
-    self.rect = self.image.get_rect()
-    self.rect.x = width*1.5
-    self.rect.y = 380 - self.image.get_height()
-
-  def update(self, speed, width):
-    self.rect.x -= speed
-    if self.rect.x < -self.width:
-      self.rect.x = width + random.randint(300, 1000)
-      self.rect.y = 150
-
-
-  def draw(self, screen):
-    screen.blit(self.image, (self.rect.x, self.rect.y))
-    print(self.rect)
